requirements.txt
fastapi==0.92.0
beanie==1.17.0
uvicorn[standard]==0.20.0
httpx==0.23.3
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.5